neutrino: fix 32<->64bit format string warnings
use portable C99 format string macros for 64bit types to
fix many warnings when compiling for 64bit architectures,
add some (int) casts for size_t

+++ src/gui/scan_setup.cpp
@@ -634,7 +634,7 @@ int CScanSetup::showScanMenuLnbSetup()
 	sat_setup->addIntroItems();
 
 	satellite_map_t & satmap = fe->getSatellites();
-	INFO("satmap size = %d", satmap.size());
+	INFO("satmap size = %d", (int)satmap.size());
 
 	CMenuWidget *tmp[satmap.size()];
 	for (sat_iterator_t sit = satmap.begin(); sit != satmap.end(); ++sit)
